+++ src/main/kotlin/com/jetbrains/snakecharm/inspections/SmkFileEndsWithCommentInspection.kt
@@ -10,6 +10,7 @@ import com.intellij.psi.PsiElement
 import com.intellij.psi.PsiFile
 import com.intellij.psi.util.nextLeaf
 import com.jetbrains.snakecharm.SnakemakeBundle
+import com.jetbrains.snakecharm.lang.psi.SmkFile
 import com.jetbrains.snakecharm.stringLanguage.lang.psi.SmkSLExpression
 
 class SmkFileEndsWithCommentInspection  : SnakemakeInspection() {
@@ -19,6 +20,10 @@ class SmkFileEndsWithCommentInspection  : SnakemakeInspection() {
         session: LocalInspectionToolSession
     ) = object : SnakemakeInspectionVisitor(holder, session) {
         override fun visitComment(comment: PsiComment) {
+            if(comment.containingFile !is SmkFile) {
+                return
+            }
+
             val nextLeaf = comment.nextLeaf(true)
             if (nextLeaf == null) {
                 registerProblem(